+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.7.0] - In Progress
+## [1.7.0] - 2019-12-26
 ### Added
 - New command: `!star` – Generates a random star sector for the *ALIEN* rpg.
 - New command: `!planet` – Generates a random uncolonized planet for the *ALIEN* rpg.